requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
jinja2==3.1.6
python-dotenv==1.2.1
httpx==0.28.1
langchain==1.2.13
langchain-core==1.2.23
langgraph==1.1.3
langchain-groq==1.1.2
